fix(frontend): stop opening Provider create wizard after SSO login
The provider-account OAuth parser treated any URL containing a bare
`code`/`error` parameter as a provider-account OAuth callback. An SSO
login callback that lands on the frontend URL (?code=&state=) was
therefore saved as a pending provider-account result, so after login
the admin console opened the Provider create wizard. Legitimate
provider-account callbacks always carry the `provider_account_oauth=1`
marker, so gate bare code/error detection on it. Manual callback paste
(allowGenericTokenNames) still accepts bare codes. Extract the parser
into a pure module with regression tests.
